Move the ws route into it's own file
